+++ app/api/getTrendingArticles/route.ts
@@ -1,0 +1,55 @@
+import { NextResponse } from "next/server";
+const trendings: Trending[] = [
+  {
+    id: 1,
+    pic: "/images/avatar.png",
+    name: "Scott Galloway 📚",
+    title: "2024 Predictions",
+    posted: "Jan 6",
+    readTime: "11 min read",
+  },
+
+  {
+    id: 2,
+    pic: "/images/avatar.png",
+    name: "Barr Moses in Towards Data Science",
+    title: "5 Hard Truth About Generative AI for Technalogy Leaders",
+    posted: "Jan 4",
+    readTime: "8 min read",
+  },
+  {
+    id: 3,
+    pic: "/images/avatar.png",
+    name: "Andrei in LlamaIndex Blog",
+    title: "A Cheat Sheet and Some Receipes For Building Advanced RAG",
+    posted: "Jan 5",
+    readTime: "7 min read",
+  },
+  {
+    id: 4,
+    pic: "/images/avatar.png",
+    name: "Ignacio de Gregorio in Towards AI",
+    title: "Apple Outclasses ChatGPT with Ferret",
+    posted: "Jan 5",
+    readTime: "7 min read",
+  },
+  {
+    id: 5,
+    pic: "/images/avatar.png",
+    name: "Lucy Foulkes",
+    title: "The adolescent mental health mess",
+    posted: "Jan 4",
+    readTime: "5 min read",
+  },
+  {
+    id: 6,
+    pic: "/images/avatar.png",
+    name: "Denilson Nastacio",
+    title: "The Ultimate Guide for Making the Best Career Choices in Tech",
+    posted: "Jan 4",
+    readTime: "8 min read",
+  },
+];
+export async function GET(request: Request) {
+  return NextResponse.json(trendings);
+}
