+++ app/loading.tsx
@@ -1,7 +1,7 @@
 export default function Loading() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <h1 className="text-3xl">Loading......</h1>
+      <span className="loading loading-spinner loading-lg"></span>
     </main>
   );
 }
