--- context/ClientThemeWrapper.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { useContext } from "react";
-import { ThemeContext } from "./ThemeContext";
-export default function ClientThemeWrapper({ children }: any) {
-  const { theme }: any = useContext(ThemeContext);
-  return <div data-theme={theme}>{children}</div>;
-}
