+++ package.json
@@ -10,10 +10,14 @@
   },
   "dependencies": {
     "next": "14.0.4",
+    "primeicons": "^6.0.1",
+    "primereact": "^10.3.1",
+    "quill": "^1.3.7",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^4.12.0",
-    "react-quill": "^2.0.0"
+    "react-markdown": "^9.0.1",
+    "rehype-raw": "^7.0.0"
   },
   "devDependencies": {
     "@types/node": "^20",
